minor changes, lots of commenting, and intialize and start populating Char_Class.py
