+++ main.py
@@ -1,8 +1,7 @@
 # from textwrap import fill
 from random import randint
 # from character import Character
-from races.Dragonborn import Dragonborn
-from Barbarian import Barbarian
+import character
 
 
 def roll_stats():
@@ -53,30 +52,36 @@ def get_class():
 
 def main():
     name = get_name()
-    # char_class = get_class()
     char_race = get_race()
     char_class = get_class()
-    if char_race == 'dragonborn':
-        a_person = Dragonborn(name=name)
 
-    if char_class == 'barbarian':
-        a_barb = Barbarian(a_person)
-    print(a_person.name)
-    print(a_person.size)
-    print(a_person.speed)
-    print(a_person.STR)
-    ages = a_person.age_range
-    print(f"You can adventure between {ages[0]} and {ages[1]} years old.")
-    print("You have the following languages:")
-    for each in a_barb.languages:
-        print(each)
+    a_person = character.Character(name=name, race=char_race, char_class=char_class)
+    print(a_person.ability_scores)
 
-    print("You have the following traits:")
-    for each in a_barb.traits:
-        print(each.replace('-', ' '))
+    # a_person.roll_ability_scores()
+    # a_person.set_ability_scores()
 
-    print("You have selected the following proficiency bonuses:")
-    print(a_barb.proficiencies)
+    # if char_race == 'dragonborn':
+    #     a_person = Dragonborn(name=name)
+    #
+    # if char_class == 'barbarian':
+    #     a_barb = Barbarian(a_person)
+    # print(a_person.name)
+    # print(a_person.size)
+    # print(a_person.speed)
+    # print(a_person.STR)
+    # ages = a_person.age_range
+    # print(f"You can adventure between {ages[0]} and {ages[1]} years old.")
+    # print("You have the following languages:")
+    # for each in a_barb.languages:
+    #     print(each)
+    #
+    # print("You have the following traits:")
+    # for each in a_barb.traits:
+    #     print(each.replace('-', ' '))
+    #
+    # print("You have selected the following proficiency bonuses:")
+    # print(a_barb.proficiencies)
 
 
 if __name__ == '__main__':
